Implement route session

+++ src/app/api/drawings/[id]/route.ts
@@ -1,11 +1,20 @@
 import prisma from "@/db";
+import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
+import authOptions from "../../auth/[...nextauth]/option";
 
 export async function DELETE(
   request: Request,
   { params }: { params: { id: string } }
 ) {
   try {
+    const session = await getServerSession(authOptions);
+    if (!session) {
+      return new NextResponse(JSON.stringify({ message: "Unauthenticated" }), {
+        status: 401,
+      });
+    }
+
     const id = Number(params.id);
 
     await prisma.drawing.delete({ where: { id } });
